Heater controller: avoid stucking at target temp. minus PID_FUNCTIONAL_RANGE
... through better initialization if I state in heater PID controller

Old behavior: always init I state with 0. This caused a threshold drop
in heating as temp. crosses target temp. minus PID_FUNCTIONAL_RANGE.
This may cause heater stuck at that temperature if heater response
is fast enough.
New behavior: initialize I state so that there is no discontinuity in
heating as a function of time -- if it is allowed by I term limits.